Factor out conftest.py.
