+++ findhelp/tests/test_models.py
@@ -1,9 +1,4 @@
-from typing import Tuple, Dict, Any, List
-from types import SimpleNamespace
-import pytest
-
 from findhelp.models import (
-    geocoding,
     to_multipolygon,
     Zipcode,
     TenantResource
@@ -13,31 +8,6 @@ from django.contrib.gis.geos import Polygon, MultiPolygon
 
 POLY_1 = Polygon.from_bbox((0, 0, 1, 1))
 POLY_2 = Polygon.from_bbox((1, 1, 2, 2))
-
-
-class FakeGeocoder:
-    def __init__(self):
-        self._registry: Dict[str, Tuple[float, float]] = {}
-
-    def register(self, address, latitude, longitude) -> None:
-        self._registry[address] = (longitude, latitude)
-
-    def search(self, address: str) -> List[Any]:
-        coords = self._registry.get(address)
-        if coords is None:
-            return []
-        result = SimpleNamespace(
-            properties=SimpleNamespace(label=address),
-            geometry=SimpleNamespace(coordinates=coords)
-        )
-        return [result]
-
-
-@pytest.fixture
-def fake_geocoder(monkeypatch):
-    fg = FakeGeocoder()
-    monkeypatch.setattr(geocoding, 'search', fg.search)
-    return fg
 
 
 def create_zipcode(zipcode='11201', geom=POLY_1):
